requirements.txt
# Core
opencv-python>=4.5.0
numpy>=1.19.0
Pillow>=8.0.0

# Deep Learning
ultralytics>=8.0.0
torch>=1.7.0
torchvision>=0.8.0

# OCR
easyocr>=1.4.1

# Web Interface
streamlit>=1.10.0

# Utils
python-dotenv>=0.19.0
scikit-learn>=0.24.0
matplotlib>=3.3.0
seaborn>=0.11.0
pandas>=1.2.0
tqdm>=4.60.0
pyyaml>=5.4.0

# For ONNX export
onnx>=1.8.0
onnxruntime>=1.7.0
